refactor(app): refactor image info hook to a shared app/odd dir
