+++ app/src/resources/dataFiles/useImageInfo.ts
